requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgi-lifespan==2.1.0
asttokens==2.4.1
asyncpg==0.29.0
attrs==23.2.0
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
dnspython==2.6.1
docopt==0.6.2
email_validator==2.1.1
executing==2.0.1
fastapi==0.111.0
fastapi-mail==1.4.1
fastapi-cli==0.0.3
fastapi-users==13.0.0
fastjsonschema==2.20.0
graphene==3.3
graphene-tornado==2.1.0
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
ipython==8.12.3
iso8601==1.1.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
makefun==1.15.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
orjson==3.10.3
passlib==1.7.4
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.2
pluggy==1.5.0
postgres==4.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
psycopg2-pool==1.2
ptyprocess==0.7.0
pure-eval==0.2.2
pwdlib==0.2.0
pycparser==2.22
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
pypika-tortoise==0.1.6
pytest==8.2.1
pytest-asyncio==0.23.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
starlette-graphene3==0.6.0
strawberry-graphql==0.235.2
tinycss2==1.3.0
tomlkit==0.12.5
tornado==6.4.1
tortoise-orm==0.20.1
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.11.0
ujson==5.9.0
urllib3==2.2.2
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==0.12.2
yarg==0.1.9